Making progress, but CMake does not make it easy, thats for sure.
