+++ VulkanExperiment/include/VulkanExperiment/experiment.h
@@ -1,0 +1,5 @@
+
+class HelloTriangleApplication {
+public:
+	void run();
+};
