--- VulkanExpermient.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29215.179
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "VulkanExpermient", "VulkanExpermient\VulkanExpermient.vcxproj", "{D1AACC31-1673-41DF-BE9F-D5939C982F3B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D1AACC31-1673-41DF-BE9F-D5939C982F3B}.Debug|x64.ActiveCfg = Debug|x64
-		{D1AACC31-1673-41DF-BE9F-D5939C982F3B}.Debug|x64.Build.0 = Debug|x64
-		{D1AACC31-1673-41DF-BE9F-D5939C982F3B}.Debug|x86.ActiveCfg = Debug|Win32
-		{D1AACC31-1673-41DF-BE9F-D5939C982F3B}.Debug|x86.Build.0 = Debug|Win32
-		{D1AACC31-1673-41DF-BE9F-D5939C982F3B}.Release|x64.ActiveCfg = Release|x64
-		{D1AACC31-1673-41DF-BE9F-D5939C982F3B}.Release|x64.Build.0 = Release|x64
-		{D1AACC31-1673-41DF-BE9F-D5939C982F3B}.Release|x86.ActiveCfg = Release|Win32
-		{D1AACC31-1673-41DF-BE9F-D5939C982F3B}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E5C3E911-5952-45CE-8364-FC55352B167C}
-	EndGlobalSection
-EndGlobal
