+++ Werld/src/main.cpp
@@ -1,0 +1,15 @@
+#include <experiment.h>
+#include <iostream>
+
+int main() {
+	HelloTriangleApplication app;
+	try {
+		app.run();
+	}
+	catch (const std::exception& e) {
+		std::cerr << e.what() << std::endl;
+		return EXIT_FAILURE;
+	}
+
+	return EXIT_SUCCESS;
+}
